+++ aubo_trajectory/trajectory_goal/trajectory_goal.cpp
@@ -1,3 +1,4 @@
+
 #include "ros/ros.h"
 #include "std_msgs/Float32MultiArray.h"
 #include "aubo_msgs/GoalPoint.h"
@@ -17,90 +18,18 @@ aubo_msgs::TraPoint traj_point;
 bool recv_flag = false;
 ros::Timer timer;
 
-aubo_msgs::TraPoint interp_one_point(aubo_msgs::TraPoint &recv_points)
-{
-    aubo_msgs::TraPoint traj_temp;
-    traj_temp.bus = recv_points.bus;
-    traj_temp.num_of_trapoint = recv_points.num_of_trapoint*2-1;
-    traj_temp.trapoints.resize(traj_temp.num_of_trapoint*6 );
-
-    for(int i = 0; i< 6;i++)
-    {
-       traj_temp.trapoints[i]=recv_points.trapoints[i];
-    }
-
-    for(int i = 1;i<traj_temp.num_of_trapoint;i++)
-    {
-       if((i%2) == 0)
-       {
-          for(int j = 0; j< 6;j++)
-          {
-            traj_temp.trapoints[i*6+j]=recv_points.trapoints[(i/2)*6+j];
-          }
-       }
-       else if((i%2) == 1)
-       {
-          for(int k = 0; k< 6;k++)
-          {
-             traj_temp.trapoints[i*6+k]= (recv_points.trapoints[(i/2)*6+k] + recv_points.trapoints[(i/2+1)*6+k])/2;
-          }
-       }
-    }
-
-    return traj_temp;
-}
-
 void chatterCallback(const aubo_msgs::TraPoint::ConstPtr &msg)
 {
-  int level = 0;
-  aubo_msgs::TraPoint trapoint_recv;
-  trapoint_recv.bus = msg->bus;
+  traj_point.bus = msg->bus;
 
-  trapoint_recv.num_of_trapoint = msg->num_of_trapoint;
-  trapoint_recv.trapoints.resize(msg->num_of_trapoint*6);
+  traj_point.num_of_trapoint = msg->num_of_trapoint;
+  traj_point.trapoints.resize(msg->num_of_trapoint*6);
 
   for(int i = 0;i<msg->num_of_trapoint;i++)
   {
      for(int j=0;j<6;j++)
      {
-       trapoint_recv.trapoints[i*6+j]=msg->trapoints[i*6+j];
-     }
-  }
-
-
-  if(msg->num_of_trapoint < 50)
-  {
-      level = 3;
-  }
-  else if(msg->num_of_trapoint < 80)
-  {
-      level = 2;
-  }
-  else if(msg->num_of_trapoint < 100)
-  {
-      level = 1;
-  }
-  else
-  {
-      level = 0;
-  }
-
-
-  if(level == 0)  traj_point = trapoint_recv;
-  else if(level >=1)
-  {
-     traj_point = interp_one_point(trapoint_recv);
-
-     if(level >= 2)
-     {
-       trapoint_recv = interp_one_point(traj_point);
-       traj_point = trapoint_recv;
-     }
-
-     if(level >= 3)
-     {
-       trapoint_recv = interp_one_point(traj_point);
-       traj_point = trapoint_recv;
+       traj_point.trapoints[i*6+j]=msg->trapoints[i*6+j];
      }
   }
 
@@ -113,6 +42,8 @@ void chatterCallback(const aubo_msgs::TraPoint::ConstPtr &msg)
 
 void chatterCallback1(const aubo_msgs::GoalPoint::ConstPtr &msg)
 {
+    ROS_INFO("goal=[%f,%f,%f,%f,%f,%f]",msg->joint1,msg->joint2,msg->joint3,msg->joint4,msg->joint5,msg->joint6);
+
     goal.bus = msg->bus;
     goal.joint1 = msg->joint1;
     goal.joint2 = msg->joint2;
@@ -133,9 +64,9 @@ void timerCallback(const ros::TimerEvent& e)
          if((i>=traj_point.num_of_trapoint)&&(traj_point.num_of_trapoint>0))
          {
             i = 0;
-           traj_point.num_of_trapoint = 0;
+            traj_point.num_of_trapoint = 0;
             recv_flag = false;
-           timer.stop();
+            timer.stop();
          }
 
          
@@ -155,7 +86,7 @@ void timerCallback(const ros::TimerEvent& e)
            else if(traj_point.bus == 1 )
            {
                cmd_pub2.publish(joints_cmd);
-               //ROS_INFO("servoj:%f,%f,%f,%f,%f,%f",joints_cmd.data[0],joints_cmd.data[1],joints_cmd.data[2],joints_cmd.data[3],joints_cmd.data[4],joints_cmd.data[5]);
+               ROS_INFO("servoj:%f,%f,%f,%f,%f,%f",joints_cmd.data[0],joints_cmd.data[1],joints_cmd.data[2],joints_cmd.data[3],joints_cmd.data[4],joints_cmd.data[5]);
            }
 
            i++;
@@ -185,7 +116,7 @@ int main(int argc, char **argv)
     ros::Subscriber sub = nh.subscribe("tra_point", 1000, chatterCallback);
     ros::Subscriber sub1 = nh.subscribe("send_goal", 1000, chatterCallback1);
 
-    timer = nh.createTimer(ros::Duration(0.008),timerCallback);
+    timer = nh.createTimer(ros::Duration(0.01),timerCallback);
     timer.stop();
   
     joints_cmd.data.resize(6);