+++ aubo_trajectory/CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog for package aubo_trajectory
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Forthcoming
+-----------
+* fixed a lib path bug
+* Update package.xml
+  add aubo_msgs to build_depend
+* Update CMakeLists.txt
+  add aubo_msgs to find_package
+* Contributors: JerryLiu, robot
+
+* fixed a lib path bug
+* Update package.xml
+  add aubo_msgs to build_depend
+* Update CMakeLists.txt
+  add aubo_msgs to find_package
+* Contributors: JerryLiu, robot
+
 0.1.1 (2016-09-20)
 ------------------
 * update CHANGELOG.rst