+++ aubo_trajectory/CHANGELOG.rst
@@ -2,8 +2,8 @@
 Changelog for package aubo_trajectory
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-Forthcoming
------------
+0.3.0 (2016-11-15)
+------------------
 * new version:add aubo_panel,aubo_new_driver
 * Contributors: robot